+++ oms/Surging.Huanlv/core/Surging.Core/Surging.Core.Domain/Entities/IAggregateRoot.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Surging.Core.Domain.Entities
 {
-    public interface IAggregateRoot : IAggregateRoot<int>, IEntity
+    public interface IAggregateRoot : IAggregateRoot<long>, IEntity
     {
     }
