+++ oms/Surging.Huanlv/core/Surging.Core/Surging.Core.Domain/Entities/IEntity.cs
@@ -2,7 +2,7 @@
 
 namespace Surging.Core.Domain.Entities
 {
-    public interface IEntity : IEntity<int>
+    public interface IEntity : IEntity<long>
     {
     }
 }
